RequirementMachine: Split off RewriteLoop.cpp from HomotopyReduction.cpp
