+++ lib/AST/CMakeLists.txt
@@ -82,6 +82,7 @@ add_swift_host_library(swiftAST STATIC
   RequirementMachine/RequirementMachine.cpp
   RequirementMachine/RequirementMachineRequests.cpp
   RequirementMachine/RewriteContext.cpp
+  RequirementMachine/RewriteLoop.cpp
   RequirementMachine/RewriteSystem.cpp
   RequirementMachine/RewriteSystemCompletion.cpp
   RequirementMachine/Symbol.cpp
